+++ packages/extension-base/src/services/earning-service/handlers/native-staking/relay-chain.ts
@@ -218,6 +218,10 @@ export default class RelayNativeStakingPoolHandler extends BaseNativeStakingPool
       } as UnstakingInfo);
     });
 
+    const tokenInfo = this.state.chainService.getAssetBySlug(this.nativeToken.slug);
+
+    await this.createWithdrawNotifications(unstakingList, tokenInfo, address);
+
     return {
       status: stakingStatus,
       balanceToken: this.nativeToken.slug,
@@ -774,7 +778,7 @@ export default class RelayNativeStakingPoolHandler extends BaseNativeStakingPool
       extrinsic = chainApi.api.tx.staking.unbond(binaryAmount);
     }
 
-    return [ExtrinsicType.STAKING_LEAVE_POOL, extrinsic];
+    return [ExtrinsicType.STAKING_UNBOND, extrinsic];
   }
 
   /* Leave pool action */