+++ packages/web-runner/src/webRunner.ts
@@ -3,9 +3,11 @@
 
 import '@subwallet/extension-inject/crossenv';
 
+import { APP_ENV, APP_VER, EnvConfig } from '@subwallet/extension-base/constants';
 import { SWHandler } from '@subwallet/extension-base/koni/background/handlers';
 import { AccountsStore } from '@subwallet/extension-base/stores';
 import KeyringStore from '@subwallet/extension-base/stores/Keyring';
+import { platformModel, platformType } from '@subwallet/extension-base/utils';
 import keyring from '@subwallet/ui-keyring';
 
 import { cryptoWaitReady } from '@polkadot/util-crypto';
@@ -23,6 +25,20 @@ setupHandlers();
 Promise.all([cryptoWaitReady(), checkRestore()])
   .then((): void => {
     console.log('[Mobile] crypto initialized');
+
+    const envConfig: EnvConfig = {
+      appConfig: {
+        environment: APP_ENV,
+        version: APP_VER
+      },
+      browserConfig: undefined,
+      osConfig: {
+        type: platformType as string,
+        version: platformModel
+      }
+    };
+
+    koniState.initEnvConfig(envConfig);
 
     // load all the keyring data
     keyring.loadAll({ store: new AccountsStore(), type: 'sr25519', password_store: new KeyringStore() });